# BigInteger parsing optimization for large decimal string