@jeffhandley 2021-07-03
@key-moon and @dotnet/area-system-numerics, I just realized that this PR was still open, but the originating branch has been deleted. I'm going to close the PR, but if you want to proceed with it, we could still carry forward starting with the [last commit](https://github.com/dotnet/runtime/pull/51953/commits/7d89c46821a9c476611419116fc8e1e0e885635a).